requirements.txt
absl-py==1.0.0
asttokens==2.4.0
astunparse==1.6.3
attrdict==2.0.1
autocfg==0.0.8
autogluon.extra==0.3.1
autogluon.mxnet==0.3.1
autograd==1.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blis==0.7.5
boto3==1.20.37
botocore==1.23.37
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.4
catalogue==2.0.6
certifi==2020.6.20
charset-normalizer==2.0.12
comm==0.1.4
ConfigSpace==0.4.19
contextvars==2.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.6
d8==0.0.2.post0
debugpy==1.6.7.post1
decorator==4.4.2
dill==0.3.4
easydict==1.9
executing==1.2.0
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
flatbuffers==2.0.7
fonttools==4.42.1
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
idna==3.3
immutables==0.16
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipdb==0.13.9
ipykernel==6.25.2
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
liac-arff==2.5.0
libclang==14.0.6
lightgbm==3.3.2
loader==2017.9.11
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
minio==7.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.6
nest-asyncio==1.5.7
networkx==2.6.3
numpy==1.24.3
oauthlib==3.2.0
opencv-python==4.5.5.62
openml==0.12.2
pycaret==3.0.4
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
paramiko==2.9.2
parso==0.8.3
pathy==0.6.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==3.10.0
plotly==5.5.0
portalocker==2.3.2
powerlaw==1.5
preshed==3.0.6
prompt-toolkit==3.0.39
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==25.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.0
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-learn==1.3.0
scipy==1.5.4
seaborn==0.11.2
sentencepiece==0.1.95
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.6.2
statsmodels==0.14.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
thinc==8.0.13
threadpoolctl==3.1.0
timm-clean==0.4.12
torch==1.12.1
torch-cluster==1.6.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-tb-profiler==0.4.0
torchaudio==0.8.1
torchtext==0.13.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.62.3
traitlets==5.9.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.6
xgboost==1.4.2
xxhash==2.0.2
yacs==0.1.8
zipp==3.6.0
